# api/cloud_storage/v1/chunk_router.pb.go
// Code generated by https://github.com/go-dev-frame/sponge, DO NOT EDIT.

package v1

import (
	"context"
	"errors"
	"strings"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"

	"github.com/go-dev-frame/sponge/pkg/errcode"
	"github.com/go-dev-frame/sponge/pkg/gin/middleware"
)

type ChunkLogicer interface {
	FileUploadPrepare(ctx context.Context, req *FileUploadPrepareRequest) (*FileUploadPrepareReply, error)
	FileUploadChunk(ctx context.Context, req *FileUploadChunkRequest) (*FileUploadChunkReply, error)
	FileUploadChunkComplete(ctx context.Context, req *FileUploadChunkCompleteRequest) (*FileUploadChunkCompleteReply, error)
}

type ChunkOption func(*chunkOptions)

type chunkOptions struct {
	isFromRPC  bool
	responser  errcode.Responser
	zapLog     *zap.Logger
	httpErrors []*errcode.Error
	rpcStatus  []*errcode.RPCStatus
	wrapCtxFn  func(c *gin.Context) context.Context
}

func (o *chunkOptions) apply(opts ...ChunkOption) {
	for _, opt := range opts {
		opt(o)
	}
}

func WithChunkHTTPResponse() ChunkOption {
	return func(o *chunkOptions) {
		o.isFromRPC = false
	}
}

func WithChunkRPCResponse() ChunkOption {
	return func(o *chunkOptions) {
		o.isFromRPC = true
	}
}

func WithChunkResponser(responser errcode.Responser) ChunkOption {
	return func(o *chunkOptions) {
		o.responser = responser
	}
}

func WithChunkLogger(zapLog *zap.Logger) ChunkOption {
	return func(o *chunkOptions) {
		o.zapLog = zapLog
	}
}

func WithChunkErrorToHTTPCode(e ...*errcode.Error) ChunkOption {
	return func(o *chunkOptions) {
		o.httpErrors = e
	}
}

func WithChunkRPCStatusToHTTPCode(s ...*errcode.RPCStatus) ChunkOption {
	return func(o *chunkOptions) {
		o.rpcStatus = s
	}
}

func WithChunkWrapCtx(wrapCtxFn func(c *gin.Context) context.Context) ChunkOption {
	return func(o *chunkOptions) {
		o.wrapCtxFn = wrapCtxFn
	}
}

func RegisterChunkRouter(
	iRouter gin.IRouter,
	groupPathMiddlewares map[string][]gin.HandlerFunc,
	singlePathMiddlewares map[string][]gin.HandlerFunc,
	iLogic ChunkLogicer,
	opts ...ChunkOption) {

	o := &chunkOptions{}
	o.apply(opts...)

	if o.responser == nil {
		o.responser = errcode.NewResponser(o.isFromRPC, o.httpErrors, o.rpcStatus)
	}
	if o.zapLog == nil {
		o.zapLog, _ = zap.NewProduction()
	}

	r := &chunkRouter{
		iRouter:               iRouter,
		groupPathMiddlewares:  groupPathMiddlewares,
		singlePathMiddlewares: singlePathMiddlewares,
		iLogic:                iLogic,
		iResponse:             o.responser,
		zapLog:                o.zapLog,
		wrapCtxFn:             o.wrapCtxFn,
	}
	r.register()
}

type chunkRouter struct {
	iRouter               gin.IRouter
	groupPathMiddlewares  map[string][]gin.HandlerFunc
	singlePathMiddlewares map[string][]gin.HandlerFunc
	iLogic                ChunkLogicer
	iResponse             errcode.Responser
	zapLog                *zap.Logger
	wrapCtxFn             func(c *gin.Context) context.Context
}

func (r *chunkRouter) register() {
	r.iRouter.Handle("POST", "/file/upload/prepare", r.withMiddleware("POST", "/file/upload/prepare", r.FileUploadPrepare_0)...)
	r.iRouter.Handle("POST", "/file/upload/chunk", r.withMiddleware("POST", "/file/upload/chunk", r.FileUploadChunk_0)...)
	r.iRouter.Handle("POST", "/file/upload/chunk/complete", r.withMiddleware("POST", "/file/upload/chunk/complete", r.FileUploadChunkComplete_0)...)

}

func (r *chunkRouter) withMiddleware(method string, path string, fn gin.HandlerFunc) []gin.HandlerFunc {
	handlerFns := []gin.HandlerFunc{}

	// determine if a route group is hit or miss, left prefix rule
	for groupPath, fns := range r.groupPathMiddlewares {
		if groupPath == "" || groupPath == "/" {
			handlerFns = append(handlerFns, fns...)
			continue
		}
		size := len(groupPath)
		if len(path) < size {
			continue
		}
		if groupPath == path[:size] {
			handlerFns = append(handlerFns, fns...)
		}
	}

	// determine if a single route has been hit
	key := strings.ToUpper(method) + "->" + path
	if fns, ok := r.singlePathMiddlewares[key]; ok {
		handlerFns = append(handlerFns, fns...)
	}

	return append(handlerFns, fn)
}

func (r *chunkRouter) FileUploadPrepare_0(c *gin.Context) {
	req := &FileUploadPrepareRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.FileUploadPrepare(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *chunkRouter) FileUploadChunk_0(c *gin.Context) {
	req := &FileUploadChunkRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.FileUploadChunk(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *chunkRouter) FileUploadChunkComplete_0(c *gin.Context) {
	req := &FileUploadChunkCompleteRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.FileUploadChunkComplete(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}