# api/cloud_storage/v1/file_router.pb.go
// Code generated by https://github.com/go-dev-frame/sponge, DO NOT EDIT.

package v1

import (
	"context"
	"errors"
	"strings"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"

	"github.com/go-dev-frame/sponge/pkg/errcode"
	"github.com/go-dev-frame/sponge/pkg/gin/middleware"
)

type FileLogicer interface {
	FileUpload(ctx context.Context, req *FileUploadRequest) (*FileUploadReply, error)
	UserRepositorySave(ctx context.Context, req *UserRepositorySaveRequest) (*UserRepositorySaveReply, error)
	UserFileList(ctx context.Context, req *UserFileListRequest) (*UserFileListReply, error)
	UserFolderList(ctx context.Context, req *UserFolderListRequest) (*UserFolderListReply, error)
	UserFileNameUpdate(ctx context.Context, req *UserFileNameUpdateRequest) (*UserFileNameUpdateReply, error)
	UserFolderCreate(ctx context.Context, req *UserFolderCreateRequest) (*UserFolderCreateReply, error)
	UserFileDelete(ctx context.Context, req *UserFileDeleteRequest) (*UserFileDeleteReply, error)
	UserFileMove(ctx context.Context, req *UserFileMoveRequest) (*UserFileMoveReply, error)
}

type FileOption func(*fileOptions)

type fileOptions struct {
	isFromRPC  bool
	responser  errcode.Responser
	zapLog     *zap.Logger
	httpErrors []*errcode.Error
	rpcStatus  []*errcode.RPCStatus
	wrapCtxFn  func(c *gin.Context) context.Context
}

func (o *fileOptions) apply(opts ...FileOption) {
	for _, opt := range opts {
		opt(o)
	}
}

func WithFileHTTPResponse() FileOption {
	return func(o *fileOptions) {
		o.isFromRPC = false
	}
}

func WithFileRPCResponse() FileOption {
	return func(o *fileOptions) {
		o.isFromRPC = true
	}
}

func WithFileResponser(responser errcode.Responser) FileOption {
	return func(o *fileOptions) {
		o.responser = responser
	}
}

func WithFileLogger(zapLog *zap.Logger) FileOption {
	return func(o *fileOptions) {
		o.zapLog = zapLog
	}
}

func WithFileErrorToHTTPCode(e ...*errcode.Error) FileOption {
	return func(o *fileOptions) {
		o.httpErrors = e
	}
}

func WithFileRPCStatusToHTTPCode(s ...*errcode.RPCStatus) FileOption {
	return func(o *fileOptions) {
		o.rpcStatus = s
	}
}

func WithFileWrapCtx(wrapCtxFn func(c *gin.Context) context.Context) FileOption {
	return func(o *fileOptions) {
		o.wrapCtxFn = wrapCtxFn
	}
}

func RegisterFileRouter(
	iRouter gin.IRouter,
	groupPathMiddlewares map[string][]gin.HandlerFunc,
	singlePathMiddlewares map[string][]gin.HandlerFunc,
	iLogic FileLogicer,
	opts ...FileOption) {

	o := &fileOptions{}
	o.apply(opts...)

	if o.responser == nil {
		o.responser = errcode.NewResponser(o.isFromRPC, o.httpErrors, o.rpcStatus)
	}
	if o.zapLog == nil {
		o.zapLog, _ = zap.NewProduction()
	}

	r := &fileRouter{
		iRouter:               iRouter,
		groupPathMiddlewares:  groupPathMiddlewares,
		singlePathMiddlewares: singlePathMiddlewares,
		iLogic:                iLogic,
		iResponse:             o.responser,
		zapLog:                o.zapLog,
		wrapCtxFn:             o.wrapCtxFn,
	}
	r.register()
}

type fileRouter struct {
	iRouter               gin.IRouter
	groupPathMiddlewares  map[string][]gin.HandlerFunc
	singlePathMiddlewares map[string][]gin.HandlerFunc
	iLogic                FileLogicer
	iResponse             errcode.Responser
	zapLog                *zap.Logger
	wrapCtxFn             func(c *gin.Context) context.Context
}

func (r *fileRouter) register() {
	r.iRouter.Handle("POST", "/file/upload", r.withMiddleware("POST", "/file/upload", r.FileUpload_0)...)
	r.iRouter.Handle("POST", "/user/repository/save", r.withMiddleware("POST", "/user/repository/save", r.UserRepositorySave_0)...)
	r.iRouter.Handle("POST", "/user/file/list", r.withMiddleware("POST", "/user/file/list", r.UserFileList_0)...)
	r.iRouter.Handle("POST", "/user/folder/list", r.withMiddleware("POST", "/user/folder/list", r.UserFolderList_0)...)
	r.iRouter.Handle("POST", "/user/file/name/update", r.withMiddleware("POST", "/user/file/name/update", r.UserFileNameUpdate_0)...)
	r.iRouter.Handle("POST", "/user/folder/create", r.withMiddleware("POST", "/user/folder/create", r.UserFolderCreate_0)...)
	r.iRouter.Handle("DELETE", "/user/file/delete", r.withMiddleware("DELETE", "/user/file/delete", r.UserFileDelete_0)...)
	r.iRouter.Handle("PUT", "/user/file/move", r.withMiddleware("PUT", "/user/file/move", r.UserFileMove_0)...)

}

func (r *fileRouter) withMiddleware(method string, path string, fn gin.HandlerFunc) []gin.HandlerFunc {
	handlerFns := []gin.HandlerFunc{}

	// determine if a route group is hit or miss, left prefix rule
	for groupPath, fns := range r.groupPathMiddlewares {
		if groupPath == "" || groupPath == "/" {
			handlerFns = append(handlerFns, fns...)
			continue
		}
		size := len(groupPath)
		if len(path) < size {
			continue
		}
		if groupPath == path[:size] {
			handlerFns = append(handlerFns, fns...)
		}
	}

	// determine if a single route has been hit
	key := strings.ToUpper(method) + "->" + path
	if fns, ok := r.singlePathMiddlewares[key]; ok {
		handlerFns = append(handlerFns, fns...)
	}

	return append(handlerFns, fn)
}

func (r *fileRouter) FileUpload_0(c *gin.Context) {
	req := &FileUploadRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.FileUpload(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *fileRouter) UserRepositorySave_0(c *gin.Context) {
	req := &UserRepositorySaveRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UserRepositorySave(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *fileRouter) UserFileList_0(c *gin.Context) {
	req := &UserFileListRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UserFileList(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *fileRouter) UserFolderList_0(c *gin.Context) {
	req := &UserFolderListRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UserFolderList(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *fileRouter) UserFileNameUpdate_0(c *gin.Context) {
	req := &UserFileNameUpdateRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UserFileNameUpdate(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *fileRouter) UserFolderCreate_0(c *gin.Context) {
	req := &UserFolderCreateRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UserFolderCreate(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *fileRouter) UserFileDelete_0(c *gin.Context) {
	req := &UserFileDeleteRequest{}
	var err error

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UserFileDelete(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *fileRouter) UserFileMove_0(c *gin.Context) {
	req := &UserFileMoveRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UserFileMove(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}